Fix SessionController

+++ lib/ptx_web/controllers/session_controller.ex
@@ -2,9 +2,15 @@ defmodule PtxWeb.SessionController do
   use PtxWeb, :controller
 
   def auth_error(conn, _params, _opts) do
-    conn
-    |> delete_resp_cookie("token")
-    |> Ptx.Guardian.Plug.sign_out()
-    |> redirect(to: conn.request_path)
+    if String.starts_with?(conn.request_path, "/api/") do
+      conn
+      |> put_status(401)
+      |> render(PtxWeb.ErrorView, :"401.json")
+    else
+      conn
+      |> delete_resp_cookie("token")
+      |> Ptx.Guardian.Plug.sign_out()
+      |> redirect(to: conn.request_path)
+    end
   end
 end
